requirements.txt
django-mptt
Django==3.0.1
Paginator==0.5.1

